# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/cart/": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get all cart",
                "responses": {}
            },
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Create cart",
                "parameters": [
                    {
                        "description": "Cart data",
                        "name": "cart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Cart"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/cart/{id}": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get cart by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "Update cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cart data",
                        "name": "cart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Cart"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Delete cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/cartitem/": {
            "get": {
                "tags": [
                    "Cart Item"
                ],
                "summary": "Get all cart item",
                "responses": {}
            },
            "post": {
                "tags": [
                    "Cart Item"
                ],
                "summary": "Create cart item",
                "parameters": [
                    {
                        "description": "CartItem data",
                        "name": "cartitem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CartItem"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/cartitem/{id}": {
            "get": {
                "tags": [
                    "Cart Item"
                ],
                "summary": "Get cart item id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "tags": [
                    "Cart Item"
                ],
                "summary": "Update cart item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CartItem data",
                        "name": "cartitem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CartItem"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "tags": [
                    "Cart Item"
                ],
                "summary": "Delete cart item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/voucher/": {
            "get": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Get ALL voucher",
                "responses": {}
            },
            "post": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Create voucher",
                "parameters": [
                    {
                        "description": "Voucher data",
                        "name": "voucher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Voucher"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/voucher/{id}": {
            "get": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Get voucher by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Update voucher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Voucher data",
                        "name": "voucher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Voucher"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Delete voucher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Cart": {
            "type": "object",
            "properties": {
                "cart_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CartItem"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "number"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.CartItem": {
            "type": "object",
            "properties": {
                "cart_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "product_id": {
                    "description": "Cart      *Cart    `json:\"cart\" gorm:\"foreignKey:CartID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE\"`",
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "cart": {
                    "$ref": "#/definitions/models.Cart"
                },
                "cart_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderDetail"
                    }
                },
                "payment_status": {
                    "description": "pending, completed, failed",
                    "type": "string"
                },
                "shipping_address": {
                    "$ref": "#/definitions/models.ShippingAddress"
                },
                "shipping_method": {
                    "$ref": "#/definitions/models.ShippingMethod"
                },
                "shipping_method_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.OrderStatus"
                },
                "status_id": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "number"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "voucher": {
                    "$ref": "#/definitions/models.Voucher"
                },
                "voucher_id": {
                    "type": "integer"
                }
            }
        },
        "models.OrderDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "order": {
                    "$ref": "#/definitions/models.Order"
                },
                "order_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "models.OrderStatus": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "description": "pending, shipped, delivered, cancelled",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "description": "Category    *Category  `json:\"category\" gorm:\"foreignKey:CategoryID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE\"` //  không cần thiết trường này",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hinh": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.ShippingAddress": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                }
            }
        },
        "models.ShippingMethod": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Order"
                    }
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "orders": {
                    "description": "Một người dùng có nhiều đơn hàng",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Order"
                    }
                },
                "password": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                }
            }
        },
        "models.Voucher": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "expiry": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Order"
                    }
                }
            }
        }
    }
}